Move "http" and "caching" packages into "io" package
