+++ src/main/java/io/github/hypixel_api_wrapper/io/caching/BasicCachingStrategy.java
@@ -1,4 +1,4 @@
-package io.github.hypixel_api_wrapper.caching;
+package io.github.hypixel_api_wrapper.io.caching;
 
 import io.github.hypixel_api_wrapper.util.Endpoint;
 import java.time.Clock;
