+++ src/main/java/io/github/hypixel_api_wrapper/io/caching/CachingStrategy.java
@@ -1,4 +1,4 @@
-package io.github.hypixel_api_wrapper.caching;
+package io.github.hypixel_api_wrapper.io.caching;
 
 import io.github.hypixel_api_wrapper.util.Endpoint;
 import org.json.JSONObject;
