+++ src/test/java/BasicCachingStrategyTest.java
@@ -3,8 +3,8 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import io.github.hypixel_api_wrapper.caching.BasicCachingStrategy;
-import io.github.hypixel_api_wrapper.caching.CachingStrategy;
+import io.github.hypixel_api_wrapper.io.caching.BasicCachingStrategy;
+import io.github.hypixel_api_wrapper.io.caching.CachingStrategy;
 import io.github.hypixel_api_wrapper.util.Endpoint;
 import java.time.Instant;
 import org.apache.commons.lang3.tuple.Pair;
